+++ crates/lintspec/src/main.rs
@@ -1,23 +1,10 @@
 #![cfg(feature = "cli")]
-use std::{collections::HashMap, env, fs::File};
+use std::env;
 
 use clap::{CommandFactory as _, Parser as _};
 use clap_complete::{generate, generate_to};
-use rayon::iter::{IntoParallelRefIterator as _, ParallelIterator};
 
-use lintspec::{
-    cli::{Args, Commands, print_reports, read_config, write_default_config},
-    error::Error,
-    files::find_sol_files,
-    lint::{ValidationOptions, lint},
-    parser::Parse as _,
-};
-
-#[cfg(feature = "slang")]
-use lintspec::parser::slang::SlangParser;
-
-#[cfg(feature = "solar")]
-use lintspec::parser::solar::SolarParser;
+use lintspec::cli::{Args, Commands, RunResult, read_config, run, write_default_config};
 
 #[allow(clippy::too_many_lines)]
 fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -54,113 +41,8 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let config = read_config(args)?;
 
-    // identify Solidity files to parse
-    let paths = find_sol_files(
-        &config.lintspec.paths,
-        &config.lintspec.exclude,
-        config.output.sort,
-    )?;
-    if paths.is_empty() {
-        return Err(String::from("no Solidity file found, nothing to analyze").into());
+    match run(&config)? {
+        RunResult::NoDiagnostics => Ok(()),
+        RunResult::SomeDiagnostics => std::process::exit(1), // indicate that there were lint diagnostics
     }
-
-    // lint all the requested Solidity files
-    let options: ValidationOptions = (&config).into();
-
-    #[allow(unused_variables)]
-    #[cfg(feature = "solar")]
-    let parser = SolarParser::default();
-
-    #[cfg(feature = "slang")]
-    let parser = SlangParser::builder()
-        .skip_version_detection(config.lintspec.skip_version_detection)
-        .build();
-
-    let diagnostics = paths
-        .par_iter()
-        .filter_map(|p| {
-            lint(
-                parser.clone(),
-                p,
-                &options,
-                !config.output.compact && !config.output.json,
-            )
-            .transpose()
-        })
-        .collect::<Result<Vec<_>, _>>()?;
-
-    // check if we should output to file or to stderr/stdout
-    let mut output_file: Box<dyn std::io::Write> = match config.output.out {
-        Some(path) => {
-            let _ = miette::set_hook(Box::new(|_| {
-                Box::new(
-                    miette::MietteHandlerOpts::new()
-                        .terminal_links(false)
-                        .unicode(false)
-                        .color(false)
-                        .build(),
-                )
-            }));
-            Box::new(
-                File::options()
-                    .truncate(true)
-                    .create(true)
-                    .write(true)
-                    .open(&path)
-                    .map_err(|err| Error::IOError {
-                        path: path.clone(),
-                        err,
-                    })?,
-            )
-        }
-        None => {
-            if diagnostics.is_empty() {
-                Box::new(std::io::stdout())
-            } else {
-                Box::new(std::io::stderr())
-            }
-        }
-    };
-
-    // no issue was found
-    if diagnostics.is_empty() {
-        if config.output.json {
-            writeln!(&mut output_file, "[]")?;
-        } else {
-            writeln!(&mut output_file, "No issue found")?;
-        }
-        return Ok(());
-    }
-
-    // some issues were found, output according to the desired format (json/text, pretty/compact)
-    if config.output.json {
-        if config.output.compact {
-            writeln!(&mut output_file, "{}", serde_json::to_string(&diagnostics)?)?;
-        } else {
-            writeln!(
-                &mut output_file,
-                "{}",
-                serde_json::to_string_pretty(&diagnostics)?
-            )?;
-        }
-    } else {
-        let cwd = dunce::canonicalize(env::current_dir()?)?;
-        let mut contents = if cfg!(any(feature = "slang", feature = "solar")) {
-            // all other clones have been dropped
-            parser.get_sources()?
-        } else {
-            HashMap::default()
-        };
-        for file_diags in diagnostics {
-            let source = contents.remove(&file_diags.document_id).unwrap_or_default();
-            print_reports(
-                &mut output_file,
-                &cwd,
-                file_diags,
-                source,
-                config.output.compact,
-            )?;
-        }
-    }
-    std::process::exit(1); // indicate that there were diagnostics (errors)
 }